Add files via upload
Cadillacs and Dinosaurs, released in Japan as Cadillacs Kyouryuu Shinseiki (キャディラックス 恐竜新世紀, Kyadirakkusu Kyōryū Shinseiki), is a 1993 arcade game by Capcom. It is a side-scrolling beat 'em up based on the comic book series Xenozoic Tales. The game was produced as a tie-in to the short-lived Cadillacs and Dinosaurs animated series which was aired during the same year the game was released. A version for the Capcom Power System Changer was planned and previewed but never released.
Up to three players can play at once, guiding their selected characters through eight stages, battling various enemies and bosses.[3] There are four playable characters, each with their own strengths and weaknesses: Jack (balanced type), Hannah (skill type), Mustapha (speed type), and Mess (power type). The game's titular dinosaurs make appearances as neutral characters that may attack both player characters and enemies.

The players have access to several attacks. Each character has two special moves including one that depletes a character's health upon contact with an enemy and when two or more players play together, they can trigger a team-attack. Players can also find and use various firearms, throwing weapons such as rocks and explosives, and melee weapons such as clubs.

When the player loses along the way, one of the big baddies holds a gun point-blank to the losing player's face in a first-person view, taunting the player with sayings such as "Eat lead, baby!". Unless the game is continued within a countdown of 20 seconds, the baddie fires his gun, leading to a game over.

Plot
The story begins in the 26th century, when an ensemble gang called the Black Marketeers begin hunting the dinosaurs to serve their unknown purpose. The continuous hunting process has made the dinosaurs violent and now they have started attacking villages and people. Four heroes: mechanic and shaman Jack Tenrec, diplomat and explorer by profession Hannah Dundee, friend and engineer Mustapha Cairo, and mysterious Mess O'Bradovich, have decided to team up against the Black Marketeers.

The protagonists journey to the "City in the Sea" where they suspect the whole hunting network being operating from. As they reach the top of a building they fight Vice Terhune. After being beaten, Vice tells them about Butcher, who had been hunting in the northern woods. Following the information, they go through the swamp forest where they find a lot of dead dinosaurs before reaching and defeating the Butcher.

Meanwhile, another big name in the hunting network, Hogg, realizes that Jack is busy in the swamp forest and decides to take over Jack's Garage. Proceeding with the mission, the heroes go through the desert of death, where they use their car to travel safely, but then the car is chased by Hogg on his cruiser motorcycle. After defeating Hogg, Jack realizes that the gangsters have taken over his garage. They go there and clear out the garage from the gangsters, eventually confronting and defeating their leader Slice and regaining control of their garage.

This is when they receive a message from an old villager who tells them about the weird behavior of the dinosaurs and asks for help. In response to the call, the heroes reach the village where they notice dinosaurs violently attacking the people and that the village has been set on fire by someone. Going ahead further, they meet the old villager again as he tells about the whole network and as soon as he is about to reveal the name of the person behind all of it, he is shot dead by Morgan, who attacks the group as well. During the fight, Morgan talks about the powers of "doctor" transforms himself into a dinosaur-like creature called Morgue. By now, they have come to know that some doctor is trying to create new lifeforms.

Proceeding ahead, the heroes reach the coal mine and another jungle, where they face a dinosaur trying to stomp them. They eventually reach a place where they fight a tentacled creature called Tyrog that attaches itself to the bodies of the gangsters as a monstrous dinosaur-human hybrid. With all the leads and hints, Jack has now realized that Dr. Simon Fessenden is the mastermind behind all that is happening.

The team heads towards his underground hideout, which again is somewhere in the "City in the Sea". The heroes head towards the bunker, going through the library and the computer lab, where the doctor appears on the computer screen, calling himself the creator of a new world. Deep down is a bio-lab, and below it is a cave, which finally leads to Fessenden's lab. Seeing the team, he transforms himself into a Morgue-like creature, but is beaten. By now, the transforming serum's effect reaches its peak, and Fessenden transforms into a three-headed creature, but the heroes manage to defeat him.

Crippled by his defeat, Fessenden sets the whole complex to self-destruct. As the laboratory starts exploding, the heroes run for their lives, but Hannah falls down while running and Jack stops to help her. Only Mustapha and Mess are able to make it out of the laboratory as it is destroyed. As they are walking back to their homes, thinking about Jack and Hannah, the latter two come from behind in the car, alive. All four heroes return home.